+++ backend/src/main/java/hanglog/category/Category.java
@@ -1,0 +1,29 @@
+package hanglog.category;
+
+import static lombok.AccessLevel.PROTECTED;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor
+public class Category {
+
+    @Id
+    private Long id;
+
+    @Column(nullable = false)
+    private String name;
+
+    @Column(nullable = false)
+    private Long googleApiId;
+
+    public Category(final String name, final Long googleApiId) {
+        this.name = name;
+        this.googleApiId = googleApiId;
+    }
+}
