+++ backend/src/main/java/hanglog/category/repository/CategoryRepository.java
@@ -1,0 +1,7 @@
+package hanglog.category.repository;
+
+import hanglog.category.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+}
