+++ backend/src/main/java/hanglog/expense/repository/ExpenseRepository.java
@@ -1,0 +1,7 @@
+package hanglog.expense.repository;
+
+import hanglog.expense.Expense;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ExpenseRepository extends JpaRepository<Expense, Long> {
+}
