+++ backend/src/main/java/hanglog/picture/domain/repository/PictureRepository.java
@@ -1,0 +1,7 @@
+package hanglog.picture.domain.repository;
+
+import hanglog.picture.domain.Picture;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PictureRepository extends JpaRepository<Picture, Long> {
+}
