+++ backend/src/main/java/hanglog/trip/domain/DayLog.java
@@ -1,0 +1,40 @@
+package hanglog.trip.domain;
+
+import static jakarta.persistence.FetchType.LAZY;
+import static jakarta.persistence.GenerationType.IDENTITY;
+import static lombok.AccessLevel.PROTECTED;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = PROTECTED)
+public class DayLog {
+
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    private Long id;
+
+    @Column(length = 50, nullable = false)
+    private String title;
+
+    @Column(nullable = false)
+    private Integer ordinal;
+
+    @ManyToOne(fetch = LAZY)
+    @JoinColumn(name = "trip_id", nullable = false)
+    private Trip trip;
+
+    public DayLog(final String title, final Integer ordinal, final Trip trip) {
+        this.title = title;
+        this.ordinal = ordinal;
+        this.trip = trip;
+    }
+}
