--- backend/src/main/java/hanglog/trip/domain/city/City.java
@@ -1,34 +1,0 @@
-package hanglog.trip.domain.city;
-
-import static jakarta.persistence.GenerationType.IDENTITY;
-import static lombok.AccessLevel.PROTECTED;
-
-import hanglog.global.BaseTimeEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import java.math.BigDecimal;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = PROTECTED)
-public class City extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private BigDecimal latitude;
-
-    @Column(nullable = false)
-    private BigDecimal longitude;
-
-    public City(final BigDecimal latitude, final BigDecimal longitude) {
-        this.latitude = latitude;
-        this.longitude = longitude;
-    }
-}
