--- backend/src/main/java/hanglog/trip/domain/expense/Expense.java
@@ -1,52 +1,0 @@
-package hanglog.trip.domain.expense;
-
-import static jakarta.persistence.FetchType.LAZY;
-import static jakarta.persistence.GenerationType.IDENTITY;
-import static lombok.AccessLevel.PROTECTED;
-
-import hanglog.trip.domain.category.Category;
-import hanglog.trip.domain.item.Item;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = PROTECTED)
-public class Expense {
-
-    @Id
-    @GeneratedValue(strategy = IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String currency;
-
-    @Column(nullable = false)
-    private Integer amount;
-
-    @OneToOne(fetch = LAZY)
-    @JoinColumn(name = "item_id", nullable = false)
-    private Item item;
-
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "category_id", nullable = false)
-    private Category category;
-
-    public Expense(final String currency,
-                   final Integer amount,
-                   final Item item,
-                   final Category category
-    ) {
-        this.currency = currency;
-        this.amount = amount;
-        this.item = item;
-        this.category = category;
-    }
-}
