--- backend/src/main/java/hanglog/trip/domain/expense/repository/ExpenseRepository.java
@@ -1,7 +1,0 @@
-package hanglog.trip.domain.expense.repository;
-
-import hanglog.trip.domain.expense.Expense;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ExpenseRepository extends JpaRepository<Expense, Long> {
-}
