--- backend/src/main/java/hanglog/trip/domain/item/type/ItemType.java
@@ -1,6 +1,0 @@
-package hanglog.trip.domain.item.type;
-
-public enum ItemType {
-    SPOT,
-    NON_SPOT
-}
