--- backend/src/main/java/hanglog/trip/domain/picture/repository/PictureRepository.java
@@ -1,7 +1,0 @@
-package hanglog.trip.domain.picture.repository;
-
-import hanglog.trip.domain.picture.Picture;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PictureRepository extends JpaRepository<Picture, Long> {
-}
