--- backend/src/main/java/hanglog/trip/domain/place/Place.java
@@ -1,63 +1,0 @@
-package hanglog.trip.domain.place;
-
-import static jakarta.persistence.FetchType.LAZY;
-import static jakarta.persistence.GenerationType.IDENTITY;
-import static lombok.AccessLevel.PROTECTED;
-
-import hanglog.trip.domain.category.Category;
-import hanglog.trip.domain.item.Item;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
-import java.math.BigDecimal;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = PROTECTED)
-public class Place {
-
-    @Id
-    @GeneratedValue(strategy = IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-    @Column(nullable = false)
-    private String address;
-
-    @Column(nullable = false)
-    private BigDecimal latitude;
-
-    @Column(nullable = false)
-    private BigDecimal longitude;
-
-    @OneToOne(fetch = LAZY)
-    @JoinColumn(name = "item_id", nullable = false)
-    private Item item;
-
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "category_id", nullable = false)
-    private Category category;
-
-    public Place(final String name,
-                 final String address,
-                 final BigDecimal latitude,
-                 final BigDecimal longitude,
-                 final Item item,
-                 final Category category
-    ) {
-        this.name = name;
-        this.address = address;
-        this.latitude = latitude;
-        this.longitude = longitude;
-        this.item = item;
-        this.category = category;
-    }
-}
