+++ backend/src/main/java/hanglog/trip/domain/repository/CityRepository.java
@@ -1,0 +1,7 @@
+package hanglog.trip.domain.repository;
+
+import hanglog.trip.domain.City;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CityRepository extends JpaRepository<City, Long> {
+}
