+++ backend/src/main/java/hanglog/trip/domain/repository/DayLogRepository.java
@@ -1,0 +1,7 @@
+package hanglog.trip.domain.repository;
+
+import hanglog.trip.domain.DayLog;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DayLogRepository extends JpaRepository<DayLog, Long> {
+}
