+++ backend/src/main/java/hanglog/trip/domain/repository/ItemRepository.java
@@ -1,0 +1,7 @@
+package hanglog.trip.domain.repository;
+
+import hanglog.trip.domain.Item;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ItemRepository extends JpaRepository<Item, Long> {
+}
