+++ backend/src/main/java/hanglog/trip/domain/repository/PlaceRepository.java
@@ -1,0 +1,7 @@
+package hanglog.trip.domain.repository;
+
+import hanglog.trip.domain.Place;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PlaceRepository extends JpaRepository<Place, Long> {
+}
