+++ backend/src/main/java/hanglog/trip/domain/repository/TripCityRepository.java
@@ -1,0 +1,7 @@
+package hanglog.trip.domain.repository;
+
+import hanglog.trip.domain.TripCity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TripCityRepository extends JpaRepository<TripCity, Long> {
+}
