--- backend/src/main/java/hanglog/trip/domain/tripcity/TripCity.java
@@ -1,39 +1,0 @@
-package hanglog.trip.domain.tripcity;
-
-import static jakarta.persistence.FetchType.LAZY;
-import static jakarta.persistence.GenerationType.IDENTITY;
-import static lombok.AccessLevel.PROTECTED;
-
-import hanglog.global.BaseTimeEntity;
-import hanglog.trip.domain.city.City;
-import hanglog.trip.domain.trip.Trip;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = PROTECTED)
-public class TripCity extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "trip_id")
-    private Trip trip;
-
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "city_id")
-    private City city;
-
-    public TripCity(final Trip trip, final City city) {
-        this.trip = trip;
-        this.city = city;
-    }
-}
