--- backend/src/main/java/hanglog/trip/domain/tripcity/repository/TripCityRepository.java
@@ -1,7 +1,0 @@
-package hanglog.trip.domain.tripcity.repository;
-
-import hanglog.trip.domain.tripcity.TripCity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TripCityRepository extends JpaRepository<TripCity, Long> {
-}
