+++ backend/src/main/java/hanglog/trip/domain/type/ItemType.java
@@ -1,0 +1,6 @@
+package hanglog.trip.domain.type;
+
+public enum ItemType {
+    SPOT,
+    NON_SPOT
+}
